Move from store to utils
